@@ -1,5 +1,5 @@
-import 'package:chatclient/utils/viewModel/viewModel.dart';
 import 'package:flutter/widgets.dart';
+import 'viewModel.dart';
 
 class ViewModelPropertyWidgetBuilder<TPropertyType>
     extends StreamBuilder<PropertyChangedEvent> {
@@ -8,12 +8,12 @@ class ViewModelPropertyWidgetBuilder<TPropertyType>
   // Methods
 
   ViewModelPropertyWidgetBuilder(
-      {Key key,
-      @required ViewModel viewModel,
-      @required String propertyName,
-      @required AsyncWidgetBuilder<PropertyChangedEvent> builder})
+      {Key? key,
+      ViewModel? viewModel,
+      required String propertyName,
+      required AsyncWidgetBuilder<PropertyChangedEvent> builder})
       : super(
             key: key,
             builder: builder,
-            stream: viewModel.whenPropertyChanged(propertyName));
+            stream: viewModel?.whenPropertyChanged(propertyName));
 }
