@@ -1,5 +1,5 @@
-import 'package:chatclient/utils/viewModel/viewModel.dart';
 import 'package:flutter/widgets.dart';
+import 'viewModel.dart';
 
 class ViewModelProvider<TViewModel extends ViewModel> extends InheritedWidget {
   // Properties
@@ -8,24 +8,24 @@ class ViewModelProvider<TViewModel extends ViewModel> extends InheritedWidget {
   // Methods
 
   ViewModelProvider(
-      {Key key,
-      @required TViewModel viewModel,
-      @required WidgetBuilder childBuilder})
+      {Key? key,
+      required TViewModel? viewModel,
+      required WidgetBuilder? childBuilder})
       : this._default(
             key: key, viewModel: viewModel, childBuilder: childBuilder);
 
   ViewModelProvider._default(
-      {Key key,
-      @required TViewModel viewModel,
-      @required WidgetBuilder childBuilder})
+      {Key? key,
+      required TViewModel? viewModel,
+      required WidgetBuilder? childBuilder})
       : assert(viewModel != null),
         assert(childBuilder != null),
-        viewModel = viewModel,
+        viewModel = viewModel!,
         super(
             key: key,
             child: ViewModelViewStateManager(
               viewModel: viewModel,
-              childBuilder: childBuilder,
+              childBuilder: childBuilder!,
             ));
 
   @override
@@ -42,11 +42,11 @@ class ViewModelViewStateManager extends StatefulWidget {
   // Methods
 
   ViewModelViewStateManager(
-      {@required ViewModel viewModel, @required WidgetBuilder childBuilder})
+      {required ViewModel? viewModel, required WidgetBuilder? childBuilder})
       : assert(viewModel != null),
         assert(childBuilder != null),
-        _childBuilder = childBuilder,
-        _viewModel = viewModel;
+        _childBuilder = childBuilder!,
+        _viewModel = viewModel!;
 
   @override
   State<StatefulWidget> createState() => _ViewModelViewStateManagerState();
