@@ -1,9 +1,9 @@
 import 'dart:async';
 
-import 'package:client/main.dart';
-import 'package:client/tests/tests.dart';
-import 'package:client/utils/viewModel/viewModel.dart';
-import 'package:client/utils/viewModel/viewModelProvider.dart';
+import '../../main.dart';
+import '../../tests/tests.dart';
+import '../../utils/viewModel/viewModel.dart';
+import '../../utils/viewModel/viewModelProvider.dart';
 import 'package:flutter/widgets.dart';
 import 'package:logging/logging.dart';
 import 'package:signalr_netcore/ihub_protocol.dart';
@@ -14,21 +14,21 @@ typedef HubConnectionProvider = Future<HubConnection> Function();
 
 class TestsPageViewModel extends ViewModel {
 // Properties
-  Logger _logger;
-  StreamSubscription<LogRecord> _logMessagesSub;
-  Tests _tests;
-  String _serverUrl;
-  HubConnection _hubConnection;
+  late Logger _logger;
+  late StreamSubscription<LogRecord> _logMessagesSub;
+  late Tests _tests;
+  late String _serverUrl;
+  HubConnection? _hubConnection;
 
-  String _errorMessage;
+  late String _errorMessage;
   static const String errorMessagePropName = "errorMessage";
   String get errorMessage => _errorMessage;
   set errorMessage(String value) {
-    updateValue(
+    updateValue<String>(
         errorMessagePropName, _errorMessage, value, (v) => _errorMessage = v);
   }
 
-  List<LogRecord> _hubLogMessages;
+  late List<LogRecord> _hubLogMessages;
   static const String hubLogMessagesPropName = "hubLogMessages";
   List<LogRecord> get hubLogMessages => _hubLogMessages;
 
@@ -76,15 +76,15 @@ class TestsPageViewModel extends ViewModel {
           .withAutomaticReconnect()
           .configureLogging(logger)
           .build();
-      _hubConnection.onclose(({error}) => _logger.info("Connection Closed"));
+      _hubConnection!.onclose(({error}) => _logger.info("Connection Closed"));
     }
 
-    if (_hubConnection.state != HubConnectionState.Connected) {
-      await _hubConnection.start();
-      _logger.info("Connection state '${_hubConnection.state}'");
+    if (_hubConnection!.state != HubConnectionState.Connected) {
+      await _hubConnection!.start();
+      _logger.info("Connection state '${_hubConnection!.state}'");
     }
 
-    return _hubConnection;
+    return _hubConnection!;
   }
 
   Future<void> connect() async {
@@ -104,12 +104,12 @@ class TestsPageViewModelProvider extends ViewModelProvider<TestsPageViewModel> {
 
   // Methods
   TestsPageViewModelProvider(
-      {Key key, viewModel = TestsPageViewModel, WidgetBuilder childBuilder})
+      {Key? key, viewModel = TestsPageViewModel, WidgetBuilder? childBuilder})
       : super(key: key, viewModel: viewModel, childBuilder: childBuilder);
 
-  static TestsPageViewModel of(BuildContext context) {
+  static TestsPageViewModel? of(BuildContext context) {
     return context
         .dependOnInheritedWidgetOfExactType<TestsPageViewModelProvider>()
-        .viewModel;
+        ?.viewModel;
   }
 }
