@@ -31,13 +31,13 @@ abstract class ViewModel {
       String propertyName,
       TPropertyType currentValue,
       TPropertyType newValue,
-      SetValue<TPropertyType> setNewValue) {
+      SetValue<TPropertyType>? setNewValue) {
     assert(setNewValue != null);
 
     if (currentValue == newValue) {
       return false;
     }
-    setNewValue(newValue);
+    setNewValue!(newValue);
     notifyPropertyChanged(propertyName);
     return true;
   }
@@ -47,34 +47,34 @@ abstract class ViewModel {
     propertyChanges.add(PropertyChangedEvent(this, propertyName));
   }
 
-  Observable<PropertyChangedEvent> whenPropertiesChanged(
-      List<String> propertyNames) {
-    assert(propertyNames != null || propertyNames.length != 0);
+  Stream<PropertyChangedEvent> whenPropertiesChanged(
+      List<String>? propertyNames) {
+    assert(propertyNames != null || propertyNames!.length != 0);
 
     return propertyChanges
         .where((event) =>
             isBlank(event.propertyName) ||
-            propertyNames.indexOf(event.propertyName) != -1)
+            propertyNames!.indexOf(event.propertyName) != -1)
         .transform(StreamTransformer.fromHandlers(handleData:
             (PropertyChangedEvent value, EventSink<PropertyChangedEvent> sink) {
       sink.add(value);
     }));
   }
 
-  Observable<void> whenPropertiesChangedHint(List<String> propertyNames) {
-    assert(propertyNames != null || propertyNames.length != 0);
+  Stream<void> whenPropertiesChangedHint(List<String>? propertyNames) {
+    assert(propertyNames != null || propertyNames!.length != 0);
 
     return propertyChanges
         .where((event) =>
             isBlank(event.propertyName) ||
-            propertyNames.indexOf(event.propertyName) != -1)
+            propertyNames!.indexOf(event.propertyName) != -1)
         .transform(StreamTransformer.fromHandlers(
             handleData: (PropertyChangedEvent value, EventSink<void> sink) {
       sink.add(null);
     }));
   }
 
-  Observable<PropertyChangedEvent> whenPropertyChanged(String propertyName) {
+  Stream<PropertyChangedEvent> whenPropertyChanged(String propertyName) {
     return propertyChanges
         .where((event) =>
             isBlank(event.propertyName) || event.propertyName == propertyName)
@@ -84,7 +84,7 @@ abstract class ViewModel {
     }));
   }
 
-  Observable whenPropertyChangedHint(String propertyName) {
+  Stream whenPropertyChangedHint(String propertyName) {
     return propertyChanges
         .where((event) =>
             isBlank(event.propertyName) || event.propertyName == propertyName)
