@@ -1,5 +1,5 @@
-import 'package:client/app.dart';
 import 'package:flutter/material.dart';
+import 'app.dart';
 
 const kServerUrl = "http://192.168.0.4:5000";
 
