@@ -1,5 +1,5 @@
-import 'package:client/tests/test.dart';
-import 'package:client/views/pages/testsPageViewModel.dart';
+import '../../views/pages/testsPageViewModel.dart';
+import '../test.dart';
 import 'package:logging/logging.dart';
 import 'package:signalr_netcore/signalr_client.dart';
 
@@ -26,9 +26,9 @@ class ServerInvokeMethodSimpleParametersNoReturnValue extends Test {
   }
 
   void _handleServerInvokeMethodSimpleParametersNoReturnValue(
-      List<Object> parameters) {
+      List<Object?>? parameters) {
     final paramValues = new StringBuffer("Parameters: ");
-    for (int i = 0; i < parameters.length; i++) {
+    for (int i = 0; i < parameters!.length; i++) {
       final value = parameters[i];
       paramValues.write("$i => $value, ");
     }
